feat: add "raw" bitwarden client that skips the json intermediate layer
