+++ crates/bitwarden-wasm/src/lib.rs
@@ -1,1 +1,3 @@
 mod client;
+mod log_level;
+mod raw_client;
