+++ crates/bitwarden-wasm/src/log_level.rs
@@ -1,0 +1,21 @@
+use log::Level;
+use wasm_bindgen::prelude::*;
+
+#[wasm_bindgen]
+pub enum LogLevel {
+    Trace,
+    Debug,
+    Info,
+    Warn,
+    Error,
+}
+
+pub fn convert_level(level: LogLevel) -> Level {
+    match level {
+        LogLevel::Trace => Level::Trace,
+        LogLevel::Debug => Level::Debug,
+        LogLevel::Info => Level::Info,
+        LogLevel::Warn => Level::Warn,
+        LogLevel::Error => Level::Error,
+    }
+}
